Add user's permissions for admin manipulations

+++ posts/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.contrib import messages
 
 # Create your views here.
@@ -8,6 +8,8 @@ from .models import Post
 from .forms import PostForm
 
 def post_create(request):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -49,6 +51,8 @@ def post_list(request): # list items
 	return render(request, 'post_list.html', context)
 
 def post_update(request, slug=None):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
@@ -65,6 +69,8 @@ def post_update(request, slug=None):
 	return render(request, 'post_form.html', context)
 
 def post_delete(request, slug=None):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, 'Successfully Deleted.')
